Unify routes names
