+++ src/routing/admin/organization/OrganizationMembersRoute.tsx
@@ -1,0 +1,20 @@
+import React, { FC, useMemo } from 'react';
+import { Route, Switch, useRouteMatch } from 'react-router-dom';
+import { FourOuFour, PageProps } from '../../../pages';
+import OrganizationCommunityPage from '../../../pages/Admin/Organization/OrganizationCommunityPage';
+
+export const OrganizationMembersRoute: FC<PageProps> = ({ paths }) => {
+  const { path, url } = useRouteMatch();
+  const currentPaths = useMemo(() => [...paths, { value: url, name: 'members', real: true }], [paths, url]);
+
+  return (
+    <Switch>
+      <Route exact path={`${path}`}>
+        <OrganizationCommunityPage paths={currentPaths} />
+      </Route>
+      <Route path="*">
+        <FourOuFour />
+      </Route>
+    </Switch>
+  );
+};
