+++ src/routing/ecoverse/EcoverseRoute.tsx
@@ -8,7 +8,7 @@ import { AuthorizationCredential } from '../../models/graphql-schema';
 import { Ecoverse as EcoversePage, FourOuFour, PageProps } from '../../pages';
 import EcoverseCommunityPage from '../../pages/community/EcoverseCommunityPage';
 import ApplyRoute from '../application/apply.route';
-import ChallengeRoute from '../challenge/challenge.route';
+import ChallengeRoute from '../challenge/ChallengeRoute';
 import RestrictedRoute, { CredentialForResource } from '../route.extensions';
 import { nameOfUrl } from '../url-params';
