+++ src/routing/messages/MessagesRoute.tsx
@@ -2,7 +2,7 @@ import React, { FC, useMemo } from 'react';
 import { Route, Switch, useRouteMatch } from 'react-router-dom';
 import { FourOuFour, Messages as MessagesPage } from '../../pages';
 
-export const Messages: FC = () => {
+export const MessagesRoute: FC = () => {
   const { path, url } = useRouteMatch();
   const currentPaths = useMemo(() => [{ value: url, name: 'messages', real: true }], []);
